hal_nxp: mcux-sdk-ng: Update drivers to SDK 25.09.00
Update drivers to SDK 25.09.00

+++ mcux/mcux-sdk-ng/drivers/lin/fsl_lin_lpuart.h
@@ -73,7 +73,7 @@ enum _lin_lpuart_flags
         (LPUART_STAT_FE_MASK), /*!< Frame error flag, sets if logic 0 was detected where stop bit expected */
     kLPUART_ParityErrorFlag = (LPUART_STAT_PF_MASK), /*!< If parity enabled, sets upon parity error detection */
 #if defined(FSL_FEATURE_LPUART_HAS_LIN_BREAK_DETECT) && FSL_FEATURE_LPUART_HAS_LIN_BREAK_DETECT
-    kLPUART_LinBreakFlag = (int)(LPUART_STAT_LBKDIF_MASK), /*!< LIN break detect interrupt flag, sets when LIN break
+    kLPUART_LinBreakFlag = (LPUART_STAT_LBKDIF_MASK), /*!< LIN break detect interrupt flag, sets when LIN break
                                                          char detected and LIN circuit enabled */
 #endif
     kLPUART_RxActiveEdgeFlag =
